Put common code between items and triggers window in a class
Should make it a little clearer in the two windows
