+++ trview.app/trview.app.vcxproj
@@ -19,6 +19,7 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="CollapsiblePanel.cpp" />
     <ClCompile Include="FileDropper.cpp" />
     <ClCompile Include="Item.cpp" />
     <ClCompile Include="ItemsWindow.cpp" />
@@ -30,6 +31,7 @@
     <ClCompile Include="WindowResizer.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="CollapsiblePanel.h" />
     <ClInclude Include="FileDropper.h" />
     <ClInclude Include="Item.h" />
     <ClInclude Include="ItemsWindow.h" />
